Move proc macro code to separate crate
