+++ Cargo.toml
@@ -9,15 +9,14 @@ categories = ["compression", "database"]
 keywords = ["compression", "timeseries", "postgres"]
 repository = "https://github.com/pganalyze/pco_store"
 
-[lib]
-proc-macro = true
-
 [dependencies]
 ahash = "0.8"
 anyhow = "1.0"
+chrono = { version = "0.4", features = ["serde"] }
 deadpool-postgres = "0.14"
 futures = "0.3"
 pco = "1.0"
+pco_store_macros = { path = "./pco_store_macros", version = "0.2.0" }
 proc-macro2 = "1.0"
 quote = "1.0"
 rmp-serde = "1.3"
@@ -30,7 +29,6 @@ tokio-postgres = "0.7"
 zstd = "0.13"
 
 [dev-dependencies]
-chrono = { version = "0.4", features = ["serde"] }
 dotenvy = "0.15"
 macrotest = "1.1"
 peak_alloc = "0.2"
