+++ pco_store_macros/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "pco_store_macros"
+version = "0.2.0"
+edition = "2024"
+license = "MIT"
+include = ["src/**/*.rs", "Cargo.toml"]
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = "2.0"
+quote = "1.0"
+proc-macro2 = "1.0"
