+++ pco_store_macros/src/decompress.rs
@@ -38,11 +38,11 @@ pub fn generate(
                 compressed_field_sizes.push(quote! { #ident.len(), });
             } else if is_nested_number(&ty) {
                 decompress_fields.push(quote! {
-                    let mut #ident: std::vec::IntoIter<#ty> = pco_decompress_nested(self.#ident)?.into_iter();
+                    let mut #ident: std::vec::IntoIter<#ty> = pco_store::nested_decompress(self.#ident)?.into_iter();
                 });
             } else {
                 decompress_fields.push(quote! {
-                    let mut #ident = serde_decompress::<#ty>(&self.#ident);
+                    let mut #ident = pco_store::serde_decompress::<#ty>(&self.#ident);
                 });
             }
             let value = if is_number(&ty) {
