+++ pco_store_macros/src/lib.rs
@@ -1,0 +1,168 @@
+use proc_macro::TokenStream;
+use proc_macro2::Span;
+use quote::{ToTokens, quote};
+use syn::parse::{Parse, ParseStream};
+use syn::{Ident, ItemStruct, Lit, Result, Token, Type, bracketed, parse_macro_input};
+
+mod decompress;
+mod fields;
+mod filter;
+mod load;
+mod new;
+mod store;
+
+#[derive(Clone)]
+struct Arguments {
+    timestamp: Option<Ident>,
+    group_by: Vec<Ident>,
+    float_round: Option<f32>,
+    table_name: Option<Ident>,
+}
+impl Parse for Arguments {
+    fn parse(input: ParseStream) -> Result<Self> {
+        let mut timestamp = None;
+        let mut group_by = Vec::new();
+        let mut float_round = None;
+        let mut table_name = None;
+        while !input.is_empty() {
+            let ident: Ident = input.parse()?;
+            let _: Token![=] = input.parse()?;
+            match ident.to_string().as_str() {
+                "timestamp" => timestamp = Some(input.parse()?),
+                "group_by" => {
+                    let content;
+                    bracketed!(content in input);
+                    group_by = content.parse_terminated(Ident::parse, Token![,])?.into_iter().collect();
+                }
+                "float_round" => {
+                    if let Lit::Int(value) = Lit::parse(input)? {
+                        let value = value.base10_parse()?;
+                        assert!(value > 0, "float_round must be greater than zero");
+                        float_round = Some(10i32.pow(value) as f32);
+                    } else {
+                        panic!("unsupported float_round value");
+                    }
+                }
+                "table_name" => table_name = Some(input.parse()?),
+                _ => {
+                    input.error("unexpected ident");
+                }
+            }
+            let _: Option<Token![,]> = input.parse().ok();
+        }
+        Ok(Self { timestamp, group_by, float_round, table_name })
+    }
+}
+
+#[proc_macro_attribute]
+pub fn store(args: TokenStream, item: TokenStream) -> TokenStream {
+    let a = args.clone();
+    let i = item.clone();
+    let args = parse_macro_input!(a as Arguments);
+    let Arguments { timestamp, group_by, float_round, table_name } = args.clone();
+    let model = parse_macro_input!(i as ItemStruct);
+    let item = proc_macro2::TokenStream::from(item);
+    let name = model.ident.clone();
+    let packed_name = Ident::new(&format!("Compressed{}s", model.ident), Span::call_site());
+
+    let table_name = if let Some(table_name) = table_name {
+        table_name.to_string()
+    } else {
+        let mut table_name = String::new();
+        for c in model.ident.to_string().chars() {
+            if c.is_uppercase() && table_name.len() > 0 {
+                table_name += "_";
+            }
+            table_name += &c.to_lowercase().to_string();
+        }
+        table_name += "s";
+        table_name
+    };
+
+    let mut packed_fields = vec![quote! { filter: Option<Filter>, }];
+    let mut timestamp_ty = None;
+    let mut using_chrono = false;
+    for field in model.fields.iter() {
+        let ident = field.ident.clone().unwrap();
+        let ty = field.ty.clone();
+        if group_by.iter().any(|i| *i == ident) {
+            packed_fields.push(quote! { #ident: #ty, });
+        } else if timestamp.as_ref().map(|t| *t == ident).unwrap_or(false) {
+            using_chrono = !ty.to_token_stream().to_string().contains("SystemTime");
+            timestamp_ty = Some(ty.clone());
+            packed_fields.push(quote! { #ident: Vec<u8>, });
+        } else {
+            packed_fields.push(quote! { #ident: Vec<u8>, });
+        }
+    }
+    if timestamp.is_some() {
+        packed_fields.push(quote! {
+            start_at: #timestamp_ty,
+            end_at: #timestamp_ty,
+        });
+    }
+    let packed_fields = tokens(packed_fields);
+
+    let filter = filter::generate(model.clone(), args.clone(), using_chrono, &timestamp_ty);
+    let fields = fields::generate(model.clone(), args.clone(), packed_name.clone());
+
+    let load_and_delete = load::generate(&model, &timestamp, &group_by, &packed_name, &table_name);
+    let decompress = decompress::generate(&model, &timestamp, &group_by, float_round, &table_name, using_chrono);
+    let store_and_store_grouped = store::generate(&model, &timestamp, &group_by, float_round, &table_name);
+    let new = new::generate(&model, &timestamp, &group_by, float_round, using_chrono);
+
+    quote! {
+        use serde::Deserialize as _;
+
+        #item
+
+        #[doc=concat!(" Generated by pco_store to store and load compressed versions of [", stringify!(#name), "]")]
+        pub struct #packed_name {
+            #packed_fields
+        }
+
+        impl #packed_name {
+            #new
+
+            #load_and_delete
+
+            #decompress
+
+            #store_and_store_grouped
+        }
+
+        #filter
+        #fields
+    }
+    .into()
+}
+
+fn is_number(ty: &Type) -> bool {
+    let ty = quote! { #ty }.to_string();
+    match ty.as_str() {
+        "u8" | "u16" | "u32" | "u64" => true,
+        "i8" | "i16" | "i32" | "i64" => true,
+        "f32" | "f64" => true,
+        "bool" => true,
+        _ => false,
+    }
+}
+
+fn is_nested_number(ty: &Type) -> bool {
+    let ty = quote! { #ty }.to_string();
+    // Remove syn's added spacing, turning "Vec < i32 >" into "Vec<i32>"
+    let ty = ty.replace(" < ", "<").replace(" >", ">");
+    match ty.as_str() {
+        "Vec<u8>" | "Vec<u16>" | "Vec<u32>" | "Vec<u64>" => true,
+        "Vec<i8>" | "Vec<i16>" | "Vec<i32>" | "Vec<i64>" => true,
+        "Vec<f32>" | "Vec<f64>" => true,
+        "Vec<bool>" => true,
+        _ => false,
+    }
+}
+
+fn tokens(input: Vec<proc_macro2::TokenStream>) -> proc_macro2::TokenStream {
+    let mut tokens = proc_macro2::TokenStream::new();
+    tokens.extend(input.into_iter());
+    tokens
+}
